[DO NOT MERGE] grus: add goodix firmware to build
This is a workaround for 2 different bugs i experienced:

1 - When compiling kernel with LTO for some weird reason Goodix firmware
is not loaded at all, touch is broken and kernel complains about the lack
of firmware files in system/etc/firmware/

2 - Compiling the kernel without LTO sometimes (it's really rare) after a
reboot touch does not work, only way to make it work is to turn off the
screen and then turn it on again. In that cases kernel log is the same
I saw in LTO kernel. The same behaviour can be randomly experienced in
recovery, which is compiled with standard kernel and cannot read firmware
in system partition because in unmounted. I'm experiencing this since the
initial release, with kernel and dt same as official los-20

Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>